+++ internal/vt10x/vt_other.go
@@ -4,11 +4,9 @@
 package vt10x
 
 import (
-	"bufio"
 	"bytes"
 	"io"
 	"unicode"
-	"unicode/utf8"
 )
 
 type terminal struct {
@@ -58,52 +56,8 @@ func (t *terminal) Write(p []byte) (int, error) {
 	return written, nil
 }
 
-// TODO: add tests for expected blocking behavior
-func (t *terminal) Parse(br *bufio.Reader) error {
-	var locked bool
-	defer func() {
-		if locked {
-			t.unlock()
-		}
-	}()
-	for {
-		c, sz, err := br.ReadRune()
-		if err != nil {
-			return err
-		}
-		if c == unicode.ReplacementChar && sz == 1 {
-			t.logln("invalid utf8 sequence")
-			break
-		}
-		if !locked {
-			t.lock()
-			locked = true
-		}
-
-		// put rune for parsing and update state
-		t.put(c)
-
-		// break if our buffer is empty, or if buffer contains an
-		// incomplete rune.
-		n := br.Buffered()
-		if n == 0 || (n < 4 && !fullRuneBuffered(br)) {
-			break
-		}
-	}
-	return nil
-}
-
-func fullRuneBuffered(br *bufio.Reader) bool {
-	n := br.Buffered()
-	buf, err := br.Peek(n)
-	if err != nil {
-		return false
-	}
-	return utf8.FullRune(buf)
-}
-
 func (t *terminal) Resize(cols, rows int) {
 	t.lock()
 	defer t.unlock()
-	_ = t.resize(cols, rows)
+	t.resize(cols, rows)
 }