+++ logger_unix_test.go
@@ -1,0 +1,5 @@
+package main
+
+import "syscall"
+
+func mkfifo(path string) error { return syscall.Mkfifo(path, 0o600) }